Extract Metamodel methods to metamodel.go
